==> 02_analysis/02_script/02_polite_speaker_ana.Rmd <==
---
title: "polite model ms - replication"
author: "Erica Yoon"
date: "May 10, 2018"
output: html_document
---

load packages.

```{r setup}
library(tidyverse); library(binom); library(langcog); library(here)
```

```{r}
d <- read.csv(file=here("02_analysis/01_data/speaker_production.csv"))
```

```{r}
ms2 <- d %>%
  filter(!is.na(positivity), !is.na(utterance)) %>% # why is there NA?
  group_by(true_state, goal) %>%
  summarise(n.total=n())

ms3 <- d %>%
  filter(!is.na(positivity), !is.na(utterance)) %>% # why is there NA?
  group_by(true_state, goal, positivity, utterance) %>%
  summarize(n = n())

ms <- left_join(ms2, ms3) %>%
  group_by(true_state, goal, positivity, utterance) %>%
  summarize(mean = n / n.total,
            ci_lower = binom.bayes(n, n.total)$lower,
            ci_upper = binom.bayes(n, n.total)$upper) %>%
  ungroup()

ms_fake <- cbind(expand.grid(true_state=levels(ms$true_state),goal=levels(ms$goal),positivity=levels(ms$positivity), utterance=levels(ms$utterance)))

ms_data <- 
  left_join(data.frame(ms_fake), data.frame(ms)) %>%
  mutate(true_state = fct_recode(true_state,
                                 "0" = "heart0", 
                                 "1" = "heart1",
                                 "2" = "heart2",
                                 "3" = "heart3"),
         utterance = fct_relevel(utterance, 
                                 "terrible", "bad", "good", "amazing"),
         goal = fct_relevel(goal, "informative", "social", "both"),
         positivity = fct_recode(positivity,
                                 "It wasn't~" = "neg",
                                 "It was~" = "no_neg"),
         positivity = fct_relevel(positivity, "It was~")) %>%
  mutate(prob = ifelse(!is.na(mean), mean, 0)) %>%
  select(-mean) %>%
  mutate(model = "NA") %>%
  mutate(source = "data")

save(ms_data, file=here("02_analysis/01_data/speaker_production.RData"))

plot_utt <- ggplot(data=ms_data,
         # mutate(positivity = fct_relevel(positivity, "not")), 
       aes(x=utterance, y=prob, color = positivity, group = positivity)) +
  geom_line()+
  facet_grid(goal~true_state, labeller=labeller(goal = label_both)) +
  xlab("") +
  ylab("proportion chosen") +
  geom_linerange(aes(ymin=ci_lower,ymax=ci_upper), position=position_dodge(width = 0.1)) +
  geom_hline(yintercept=.125, lty=2) +
  ylim(0,1)+
  scale_color_solarized()+
  ggthemes::theme_few()+
  labs(color="") +
  theme(axis.text.x = element_text(angle = 45, vjust=0.5),
        legend.position = "bottom")
plot_utt

ggsave("speaker_production_utt.png", plot = plot_utt, width = 7, height = 5,
       path = here::here("02_analysis/03_figs"))


# ms.all.4pred <- ms.all

```

```{r}
ms2 <- d %>%
  filter(!is.na(positivity), !is.na(utterance)) %>% # why is there NA?
  group_by(true_state, goal) %>%
  summarise(n.total=n())

ms3 <- d %>%
  filter(!is.na(positivity), !is.na(utterance)) %>% # why is there NA?
  group_by(true_state, goal, positivity) %>%
  summarize(n = n())

ms <- left_join(ms2, ms3) %>%
  group_by(true_state, goal, positivity) %>%
  summarize(mean = n / n.total,
            ci_lower = binom.bayes(n, n.total)$lower,
            ci_upper = binom.bayes(n, n.total)$upper) 

ms_fake <- cbind(expand.grid(true_state=levels(ms$true_state),goal=levels(ms$goal),positivity=levels(ms$positivity)), mean=NA, ci_lower=NA, ci_upper=NA)
  
ms.all <- rbind(data.frame(ms), data.frame(ms_fake)) %>%
  mutate(true_state = fct_recode(true_state,
                                 "0" = "heart0", 
                                 "1" = "heart1",
                                 "2" = "heart2",
                                 "3" = "heart3"),
         goal = fct_relevel(goal, "informative", "social", "both"),
         positivity = fct_recode(positivity,
                                 "not" = "neg",
                                 "yes" = "no_neg"),
         positivity = fct_relevel(positivity, "yes")) %>%
  rename(prob = mean) %>%
  mutate(model = "NA", source = "data") %>%
  filter(!is.na(prob), positivity == "not") %>%
  select(-positivity)

ms_data_neg <- ms.all

save(ms_data_neg, file=here("02_analysis/01_data/speaker_production_neg.RData"))
```

```{r}
speaker_production_neg <- ms.all %>%  
  filter(!is.na(mean)) %>%
  # filter(positivity == "not") %>%
ggplot(., 
         # filter(positivity == "not"),
       aes(x=true_state, y=prob, 
           # linetype = positivity, 
           color=goal, 
           group=goal)) +
           # group=interaction(goal, positivity))) +
  geom_line(stat="identity", position=position_dodge()) +
  xlab("true state") +
  ylab("proportion negation") +
  geom_linerange(aes(ymin=ci_lower,ymax=ci_upper), position=position_dodge(width=.05)) +
  # scale_colour_discrete(guide = guide_legend(title = "utterance type")) +
  theme_few()+
  scale_color_solarized()
  # ggtitle("Speaker production: Compare neg (\"not\") vs. no neg (\"yes\")")

ggsave("speaker_production_neg.png", plot = speaker_production_neg, width = 5.5, height = 3,
       path = here::here("02_analysis/03_figs"))
```

# literal semantics

```{r, fig.width = 10, fig.height = 4}
d <- read.csv(here("02_analysis/01_data/literal_semantics.csv")) %>%
  mutate(utterance = fct_relevel(utterance, "terrible", "bad", "good", "amazing"),
         positivity = fct_recode(positivity, 
                                 "It wasn't~" = "it wasn't ___",
                                 "It was~" = "it was ___")
                                 )

ms <- d %>%
  group_by(positivity, state, utterance, subid) %>%
  summarize(
            judgment = mean(judgment, na.rm=TRUE)
          ) %>%
  group_by(positivity, state, utterance) %>%
  multi_boot_standard(col = "judgment") %>%
  mutate(judgment = mean)
```

```{r echo = FALSE, fig.width = 12, fig.height = 4}
qplot(state, judgment, 
      colour = positivity,
      data=ms) + 
  geom_line(aes(group=positivity)) +
  facet_grid(.~utterance) +
  xlab("state (0=worst)") +
  ylab("proportion of\n acceptances") +
  # ggtitle("Literal semantics") +
  geom_errorbar(aes(ymin=ci_lower,ymax=ci_upper,width=.1)) +
  # scale_color_discrete(guide_legend(title="")) +
  theme_few() +
  
  scale_color_solarized() +
  theme(legend.position = "bottom", 
        legend.title = element_blank()) +
  geom_hline(yintercept = .5, lty=2)
      # theme(axis.text=element_text(size=14),
      #   axis.title=element_text(size=14,face="bold"),
      #   strip.text.x=element_text(size=14,face="bold"),
      #   strip.text.y=element_text(size=14,face="bold"),
      #   legend.text=element_text(size=13))


# ms
```

# stats

## lmer

We do not report our lmer values, but we provide the analysis below.

```{r}
# lmer to look at neg ~ state * goal
library(lme4)
d <- read.csv(file=here("02_analysis/01_data/speaker_production.csv"))

ms_glmer <- d %>%
  mutate(positivity = factor(positivity, labels = c(0,1))) %>%
  mutate(positivity = as.numeric(as.character(positivity))) %>%
  mutate(true_state = as.numeric(substr(true_state, 6, 6)))
summary(glmer(data=ms_glmer, positivity ~ true_state * goal * (1 |subid), family=binomial)) # fuller random effect structure causes model convergence failure
```

## brms

```{r}
# brms

# brms_pol <- brm(data=ms_glmer %>%
#                   mutate(goal = relevel(goal, ref = "both")) %>%
#                   mutate(true_state = scale(true_state)),
#                 positivity ~ true_state * goal + (true_state + goal |subid) + (true_state + goal | item), family="bernoulli")

load(here("03_model/03_other_stat", "brms_polite.Rds"))

summary(brms_pol)
```
