==> 02_analysis/02_script/03_understanding-prsa.Rmd <==
---
title: "Politeness S2 predictions"
output: html_notebook
---


```{r setup, include=FALSE}
library(tidyverse)
library(rwebppl)
library(jsonlite)
library(magrittr)
library(ggthemes)
library(forcats)
library(knitr)
library(gganimate)
library(RColorBrewer)
library(cowplot)
library(here)
library(viridis)

```


```{r languageUtils}
language <- '
var cost_yes = 1;
var cost_neg = 2;

var isNegation = function(utt){
  return (utt.split("_")[0] == "not")
};

var utterances = [
  "yes_terrible","yes_bad","yes_good","yes_amazing",
  "not_terrible","not_bad","not_good","not_amazing"
  ];

var uttCosts = map(function(u) {
	return isNegation(u) ? Math.exp(-cost_neg) : Math.exp(-cost_yes)
}, utterances)

var utterancePrior = Infer({method: "enumerate"}, function(){
  return utterances[discrete(uttCosts)]
});

var literalSemantics = {
  "state": [0, 1, 2, 3],
  "not_terrible": [0.001, 0.353, 0.980, 0.863],
  "not_bad": [0.020, 0.157, 0.999, 0.902],
  "not_good": [0.961, 0.863, 0.039, 0.020],
  "not_amazing": [0.961, 0.980, 0.627, 0.039],
  "yes_terrible": [0.961, 0.627, 0.039, 0.039],
  "yes_bad": [0.980, 0.882, 0.039, 0.020],
  "yes_good": [0.001, 0.020, 0.941, 0.999],
  "yes_amazing": [0.001, 0.001, 0.216, 0.980]
};

var meaning = function(words, state){
  return flip(literalSemantics[words][state]);
};
'
```

```{r utils}
utils <- '
var states = [0, 1, 2, 3];

var statePrior = function(){
  return uniformDraw(states);
};

var speakerOptimality = 2.5;

var round = function(x){
	return Math.round(x * 100) / 100
}

// var phiWeights = [1, 1, 1, 1, 1, 1, 1, 1, 1]
// var weightBins = [0.1, 0.3, 0.5, 0.7, 0.9]
// var weightBins = [0.01, 0.3, 0.5, 0.7, 0.99]
 var weightBins = [0.05, 0.25, 0.5, 0.75, 0.95]

// var weightBins = map(round, _.range(0.05,1, 0.05))
// var weightBins = map(round, _.range(0.1,1, 0.1))
var phiWeights = repeat(weightBins.length, function(){1})

var StatePrior = Categorical({vs: states, ps:[1,1,1,1]})
'
```



```{r}
prsa <-'
var listener0 = cache(function(utterance) {
  Infer({method: "enumerate"}, function(){
    var state = sample(StatePrior)

    var m = meaning(utterance, state);
    condition(m);
    return state;
	})
}, 10000);

var speaker1 = cache(function(state, phi) {
  Infer({method: "enumerate"}, function(){

    var utterance = sample(utterancePrior);
    var L0 = listener0(utterance);

    var epistemicUtility = L0.score(state);
    var socialUtility = expectation(L0, function(s){return s});
    var eUtility = phi*epistemicUtility;
    var sUtility = (1-phi)*socialUtility;
    var speakerUtility = eUtility+sUtility;

    factor(speakerOptimality*speakerUtility);

    return utterance;
	})
}, 10000);

var listener1 = cache(function(utterance) {
  Infer({method: "enumerate"}, function(){

   var phi = categorical ({vs: weightBins, ps: phiWeights})
   var state = sample(StatePrior)

   var S1 = speaker1(state, phi)
   observe(S1, utterance)

   return { state, phi }

 })
}, 10000);

  var speaker2 = cache(function(state, S1_phi, selfpres_phi, social_phi, informative_phi) {
    Enumerate(function(){
    var utterance = sample(utterancePrior);
    var normalized_phis = normalize([selfpres_phi, social_phi, informative_phi])
    var phi_selfpres = normalized_phis[0];
    var phi_social = normalized_phis[1];
    var phi_informative = normalized_phis[2];

    var L1 = listener1(utterance)
	  var L1_goal = marginalize(L1, "phi");
	  var L1_state = marginalize(L1, "state");

    var epistemicUtility = L1_state.score(state);
    var selfpresUtility = L1_goal.score(S1_phi);
    var socialUtility = expectation(L1_state, function(s){return s});
    var speakerMixedUtility = phi_informative*epistemicUtility +
                              phi_selfpres*selfpresUtility +
                              phi_social*socialUtility

	factor(speakerOptimality * speakerMixedUtility);

      return utterance

    })
  }, 10000)
'
```

```{r}
literalListenerCall <- '
_.fromPairs(_.flatten(map(function(u){
  var post = listener0(u)
  var stateObj = map(function(s){
    [u + "_" + s, Math.exp(post.score(s))]
  }, states)
  return stateObj
}, utterances)))
'
```

```{r}
pragmaticListenerCall <- '
_.fromPairs(_.flatten(map(function(u){
  var post = listener1(u)
  var stateObj = _.flatten(map(function(phi){
      map(function(s){
            [u + "_" + s + "_" + phi*100, Math.exp(post.score({state: s, phi: phi}))]
          }, states)
    }, weightBins))
  return stateObj
}, utterances)))
'
```


```{r}
speakerCall <- '
_.flatten(
map(function(phi){
    map(function(s){
      var speakProbs = speaker1(s, phi)
      var uttObj = _.fromPairs(map(function(u){
        [u, Math.exp(speakProbs.score(u))]
      }, utterances))
       return extend(uttObj,{ state: s, phi: phi, })
    }, states)
}, weightBins))
'
```

```{r}
pragmaticSpeakerCall <- '
_.flattenDeep(
map(function(S1_phi){
    map(function(selfpres_phi){
        map(function(social_phi){
            map(function(informative_phi){
                map(function(s){
                   display(s + " " + S1_phi + " " +  selfpres_phi+ " " +  social_phi + " " +  informative_phi)
                  var speakProbs = speaker2(s, S1_phi, selfpres_phi, social_phi, informative_phi)
                  var uttObj = _.fromPairs(map(function(u){
                    [u, Math.exp(speakProbs.score(u))]
                  }, utterances))
                   return extend(uttObj,{ state: s, phi_S1: S1_phi, 
                                          phi_selfpres: normalize([selfpres_phi, social_phi, informative_phi])[0], 
                                          phi_social: normalize([selfpres_phi, social_phi, informative_phi])[1], 
                                          phi_informative: normalize([selfpres_phi, social_phi, informative_phi])[2]})
                }, states)
            }, weightBins)
        }, weightBins)
    }, weightBins)
}, weightBins)
)
'
```

### L0 viz

```{r}
rs.l0 <- webppl(paste(utils, language, prsa, literalListenerCall,  sep = '\n'))


rs.l0.tidy <- data.frame(rs.l0) %>% gather(utt_state, prob) %>%
  separate(utt_state, into=c("positive", "utt", "state")) %>% 
    mutate(utterance = paste(positive, utt),
                utterance = factor(utterance,
                                   levels = c("yes terrible",
                                              "yes bad",
                                              "yes good",
                                              "yes amazing",
                                              "not terrible",
                                              "not bad",
                                              "not good",
                                              "not amazing")))
```

#### L0 static

```{r}
fig.l0 <- ggplot(rs.l0.tidy %>%
                   mutate(positive = 
                          factor(positive, levels = c("yes", "not")),
                          positive = fct_recode(positive, 
                                               "It was ~" = "yes",
                                               "It wasn't ~" = "not"),
                          utt = factor(utt, 
                                       levels = c("terrible", 
                                                  "bad",
                                                  "good", 
                                                  "amazing"))),
       aes( x = state, y = prob, group = state, fill = positive) )+
  geom_bar(position = position_dodge(), stat = 'identity', color = 'black')+
  scale_y_continuous(limits = c(0, 0.9), breaks = c(0, 0.25, 0.5, 0.75))+
  facet_grid(positive~utt)+
  # scale_fill_manual(values =  rev(brewer.pal(5,"RdBu")))+
  # scale_fill_manual(values =  brewer.pal(4,"RdBu"))+
  scale_fill_solarized()+
  guides(fill = F)+
  theme_few()+
  ylab("Literal listener probability")

fig.l0
# ggsave("img/l0.pdf", width = 7, height = 4)
```

```{r ms-fig-l0}
fig.l0.heat <- ggplot(rs.l0.tidy %>% 
                                           mutate(utt = fct_recode(utterance, 
                                            "amazing" = "yes amazing", 
                                            "good" = "yes good", 
                                            "bad" = "yes bad", 
                                            "terrible" = "yes terrible"),
                          utt = fct_relevel(utt, "terrible", "bad", "good", "amazing", "not terrible", "not bad", "not good", "not amazing"),
                        positive = fct_relevel(positive, "yes")),               
                 aes ( x = state, y = utt, fill = prob))+
  geom_tile(colour = "grey95") + 
  scale_fill_gradient(low = "white", high = "seagreen")+
  facet_grid(positive ~ ., switch="y", scales = "free_y")+
  guides(fill = F)+
  ylab("utterance") +
  xlab("true state") +
  ggtitle(expression(L[0])) +
  theme_few() +
  theme(
        strip.text.y = element_blank()
        )  

fig.l0.heat
ggsave("../../04_manuscript/01_general/fig/l0_schematic.png", width = 3.5, height = 4)
```

### S1 viz

```{r}
rs.s1 <- webppl(paste(utils, language, prsa, speakerCall,  sep = '\n'))


rs.s1.tidy <- rs.s1 %>%
  gather(utt, prob, -state, -phi) %>%
                   mutate(utt2 = utt) %>%
                   mutate(utt = fct_recode(utt, "not amazing" = "not_amazing", 
                                            "not terrible" = "not_terrible", 
                                            "not good" = "not_good", 
                                            "not bad" = "not_bad", 
                                            "amazing" = "yes_amazing", 
                                            "good" = "yes_good", 
                                            "bad" = "yes_bad", 
                                            "terrible" = "yes_terrible"),
                          utt = fct_relevel(utt, "terrible", "bad", "good", "amazing", "not terrible", "not bad", "not good", "not amazing")) %>%
                            
                   separate(utt2, c("positive", "word")) %>%
                   mutate(word = fct_relevel(word, "terrible", "bad", "good", "amazing")) %>%
                   mutate(positive = fct_relevel(positive, "yes")) %>%
                   mutate(positive = fct_recode(positive, "it was ~" = "yes", "it wasn't ~" = "not")) %>%
  mutate(goal = case_when(phi == 0.95 ~ "informational", 
                        phi == 0.05 ~ "social"
                        # (phi == 0.25) ~ "both"
                        )) %>%
                   # mutate(goal = fct_relevel(goal, "informational", "social", "both")) %>%
                   mutate(goal = fct_relevel(goal, "informational", "social", "both")) %>%
                   filter(!is.na(goal))
```

```{r ms-fig-s1}
fig.s1 <- ggplot(rs.s1.tidy, 
                 aes ( x = state, y = utt, fill = prob))+
  geom_tile(colour = "grey95") + 
  # scale_fill_viridis(option="plasma") +
  scale_fill_gradient(low = "white", high = "orange")+
  facet_grid(positive ~ goal, switch="y", scales = "free_y")+
  # scale_y_continuous(limits = c(0, 1), breaks = c(0, 0.5, 1))+
  guides(fill = F)+
  ylab("utterance") +
  # scale_y_continuous(position = "right") +
  # scale_x_discrete(position = "top") +

  xlab("true state") +
  ggtitle(expression(S[1])) +
  
  # theme_bw()+
  theme_few() +
  theme(
    # axis.text.x = element_text(angle = 45, hjust=1),
        # strip.text.x = element_text(angle = 180),
        strip.text.y = element_blank()
        )  
# theme(strip.placement = "outside")
  # theme(        panel.border = element_rect(colour = "black", fill=NA, size=1))
  # panel_border(remove = TRUE)
  
fig.s1
ggsave("../../04_manuscript/01_general/fig/s1_utility_schematic_stateBottom.png", width = 5, height = 4)
```


#### S1 movie

```{r}
mov.s1 <- rs.s1.tidy %>%
  separate(utt, into = c("positive", "utterance")) %>%
  mutate(positive = factor(positive, levels = c("not", "yes")),
         utterance = factor(utterance, levels = c("terrible", "bad", "good", "amazing")),
         phi = 1 - phi) %>%
  filter(state %in% c(1, 3, 5))%>%
  ggplot(., aes( x = utterance, fill = positive, y = prob ))+
  geom_bar(stat = 'identity', position = position_dodge(), color = 'black', aes(frame = phi))+
  facet_wrap(~state, nrow = 1)+
  scale_fill_solarized()+
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1))+
  scale_y_continuous(limits = c(0, 0.6), breaks = c(0, 0.25, 0.5))+
  ylab("Speaker production probability")+
  ggtitle("relative social utility = ")

mov.s1
# gganimate(mov.s1, "img/s1.gif", ani.width=600, ani.height=250, interval = 0.3)
```

#### S1 static

```{r}
fig.s1 <- rs.s1.tidy %>%
  filter(phi %in% c(0.10, 0.30, 0.50, 0.70)) %>%
  separate(utt, into = c("positive", "utterance")) %>%
  mutate(positive = factor(positive, levels = c("yes", "not")),
         positive = fct_recode(positive, 
                              "It was ~" = "yes",
                              "It wasn't ~" = "not"),
         utterance = factor(utterance, levels = c("terrible", "bad", "good", "amazing")),
         phi = 1 - phi) %>%
  ggplot(., aes( x = utterance, fill = positive, y = prob))+
  geom_bar(stat = 'identity', position = position_dodge(), color = 'black')+
  facet_grid(phi~state, labeller = label_both)+
  scale_fill_solarized()+
  theme_few()+
  theme(legend.position="top",
        legend.title=element_blank(),
        axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1),
        strip.text.y = element_text(angle = 0))+
  scale_y_continuous(limits = c(0, 1), breaks = c(0, 0.25, 0.5, 0.75, 1))+
  ylab("Speaker production probability")
fig.s1
# ggsave("img/s1.pdf", width = 8, height = 4)
```


### L1 viz

```{r}
rs.l1 <- webppl(paste(utils, language, prsa, pragmaticListenerCall,  sep = '\n'))

rs.l1.tidy <- data.frame(rs.l1) %>% gather(utt_state, prob) %>%
  separate(utt_state, into = c("positive", "utt", "state", "phi"), sep  = "_") %>%
  mutate(positive2 = positive, utt2 = utt) %>%
  unite("utterance", c("positive2", "utt2")) %>%
  mutate(positive = factor(positive, levels = c("yes", "not")),
         # positive = fct_recode(positive, 
         #                       "it was ~" = "yes",
         #                       "it wasn't ~" = "not"),
         utt = factor(utt, levels = c(
           "terrible",
           "bad",
           "good",
           "amazing"
         )),
         fi = as.numeric(phi) / 100,
         # phi = fct_recode(phi, 
         #                  "phi1[S[1]] = 0.05" = "0.05",
         #                  "phi1[S[1]] = 0.25" = "0.25",
         #                  "phi1[S[1]] = 0.5" = "0.5",
         #                  "phi1[S[1]] = 0.75" = "0.75",
         #                  "phi1[S[1]] = 0.95" = "0.95"
         #                  ),
         utt = fct_recode(utterance, "not amazing" = "not_amazing", 
                                            "not terrible" = "not_terrible", 
                                            "not good" = "not_good", 
                                            "not bad" = "not_bad", 
                                            "amazing" = "yes_amazing", 
                                            "good" = "yes_good", 
                                            "bad" = "yes_bad", 
                                            "terrible" = "yes_terrible"),
                          utt = fct_relevel(utt, "terrible", "bad", "good", "amazing", "not terrible", "not bad", "not good", "not amazing"))
```

joint posterior

```{r ms-fig-l1}
fig.l1 <- ggplot(rs.l1.tidy, 
                 # aes ( x = state, y = utt, fill = prob)
                 aes ( x = state, y = factor(fi), fill = prob)
                 )+
  geom_tile(colour = "grey95") + 
  scale_fill_gradient(low = "white", high = "slateblue1")+
  scale_y_discrete(expand=c(0.01, 0.01)) +
  # facet_grid(positive*utt ~ fi, switch="y", scales = "free_y", labeller = label_bquote(cols = phi1[S[1]] == .(as.numeric(fi))))+
  facet_wrap(. ~ utt, ncol=4)+
  # scale_y_continuous(limits = c(0, 1), breaks = c(0, 0.25, .5, .75, 1))+
  guides(fill = F)+
  ylab("speaker's relative informational utility \ninferred by pragmatic listener") +
    # ylab("inferred phi1[S[1]]") +
theme_few()+
  ggtitle(expression(L[1])) +
  theme(
        strip.text.y = element_blank(),
      panel.spacing=unit(0, "lines")
      # panel.spacing.y=unit(2, "lines")
# panel.border = element_blank()
        )

  # panel_border(remove = TRUE)
fig.l1
# ggsave("../../04_manuscript/01_general/fig/l1_schematic.png", width = 8, height = 4)
ggsave("../../04_manuscript/01_general/fig/l1_schematic_v2.png", width = 8, height = 4)
```

marginal posterior on phi

```{r}
rs.l1.tidy.phi <- rs.l1.tidy %>%
  group_by(positive, utt, phi) %>%
  summarize(prob = sum(prob))


fig.marginal.phi.l1 <- ggplot(rs.l1.tidy.phi %>%
         filter(utt %in% c("terrible", "amazing")),
       aes( x = phi, y = prob, group = phi, fill = positive) )+
  geom_bar(position = position_dodge(), stat = 'identity', color = 'black',
           alpha =0.8)+
  scale_y_continuous(limits = c(0, 0.11), breaks = c(0, 0.1))+
  scale_x_continuous(limits = c(0, 1), breaks = c(0,  1))+#,
                     #labels = c("epistemic", "social"))+
  facet_grid(positive~utt, scales = 'free')+
  # scale_fill_manual(values =  rev(brewer.pal(5,"RdBu")))+
  # scale_fill_manual(values =  brewer.pal(4,"RdBu"))+
  scale_fill_ptol(name="") +
  # scale_fill_solarized(name="")+
  # guides(fill = F)+
  theme_few()+
  theme(#axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1),
        legend.position="top",
        strip.text.y = element_blank(),
         strip.text.x = element_text(size=14),
       # axis.text.x = element_text(hjust=-2),
        # strip.text.y = element_text(angle = 0),
         legend.text = element_text(size=15),
       text = element_text(size = 14.5)) +

  ylab("probability mass")+
  xlab(expression("inferred " * S[1] * "'s relative informational utility (" * phi1[S[1]] * ")"))

fig.marginal.phi.l1
```

```{r}
rs.l1.tidy.state <- rs.l1.tidy %>%
  group_by(positive, utt, state) %>%
  summarize(prob = sum(prob))

fig.marginal.state.l1 <- ggplot(rs.l1.tidy.state %>%
         filter(utt %in% c("terrible", "amazing")),
       aes( x = state, y = prob, group = state, fill = positive) )+
  geom_bar(position = position_dodge(), stat = 'identity', color = 'black')+
  scale_y_continuous(limits = c(0, 0.7), breaks = c(0, 0.3, 0.6))+
  #scale_x_continuous(limits = c(0, 1), breaks = c(0,  1))+#,
                     #labels = c("epistemic", "social"))+
  facet_grid(positive~utt, scales = 'free')+
  # scale_fill_manual(values =  rev(brewer.pal(5,"RdBu")))+
  # scale_fill_manual(values =  brewer.pal(4,"RdBu"))+
  scale_fill_ptol() +
  # scale_fill_solarized()+
  
  guides(fill = F)+
  theme_few()+
  theme(#axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1),
        strip.text.x = element_text(size=14),
        strip.text.y = element_blank(),
        text = element_text(size = 14.5)
        )+
  
  ylab("probability mass")+
  xlab("inferred state (number of hearts)")

fig.marginal.state.l1
```





```{r}
fig.l1 <- ggplot(rs.l1.tidy,
       aes( x = state, y = prob, group = state, fill = positive) )+
  geom_bar(position = position_dodge(), stat = 'identity', color = 'black')+
  scale_y_continuous(limits = c(0, 0.6), breaks = c(0, 0.25, 0.5))+
  facet_grid(positive~utt)+
  # scale_fill_manual(values =  rev(brewer.pal(5,"RdBu")))+
  # scale_fill_manual(values =  brewer.pal(4,"RdBu"))+
  scale_fill_solarized()+
  guides(fill = F)+
  theme_few()+
  ylab("Pragmatic listener probability")
fig.l1
```

### S2 viz

```{r}
# schematic
# rs.s2 <- webppl(paste(utils, language, prsa, pragmaticSpeakerCall,  sep = '\n'))
# save(rs.s2, file=here("02_analysis/01_data/s2_schematic_predictions_v2.RData"))
load(here("02_analysis/01_data/s2_schematic_predictions_v2.RData"))

# full model
# load(here("model/data", "bda-utterance-allMods.RData"))

# data
# load(here("model/data",a "speakerProduction.RData"))

# inferred full model weights
# load(here("model/data", "bda-goalWeights.RData"))

rs.s2.tidy <- rs.s2 %>%
  gather(utt, prob, -state, -phi_S1, -phi_informative, -phi_social, -phi_selfpres) %>% 
  mutate(phi_selfpres_bin = as.numeric(as.character(cut(phi_selfpres, 
                                                        breaks = c(0, 0.2, 0.4, 0.6, 0.8, 1), 
                                                        labels = c(0.1, 0.3, 0.5, 0.7, 0.9)))),
         phi_informative_bin = as.numeric(as.character(cut(phi_informative, 
                                                           breaks = c(0, 0.2, 0.4, 0.6, 0.8, 1), 
                                                           labels = c(0.1, 0.3, 0.5, 0.7, 0.9)))),
         phi_social_bin = as.numeric(as.character(cut(phi_social, 
                                                      breaks = c(0, 0.2, 0.4, 0.6, 0.8, 1), 
                                                      labels = c(0.1, 0.3, 0.5, 0.7, 0.9))))) %>%
  mutate( 
    goal = as.factor(case_when(
      phi_informative_bin == 0.9 & phi_social_bin == 0.1 & phi_selfpres_bin == 0.1 ~ "be truly informative",  
      phi_informative_bin == 0.1 & phi_social_bin == 0.9 & phi_selfpres_bin == 0.1 ~ "be truly social",
      # (phi_informative_bin == 0.5 | phi_informative_bin == 0.3 | phi_informative_bin == 0.7) & (phi_social_bin == 0.5 | phi_social_bin == 0.7 | phi_social_bin == 0.3) & phi_selfpres_bin == 0.1 ~ "be truly both",
      phi_informative_bin == 0.5 & phi_social_bin == 0.5 & phi_selfpres_bin == 0.1 & phi_S1 == 0.1  ~ "be truly both", 
      phi_informative_bin == 0.1 & phi_social_bin == 0.1 & (phi_selfpres_bin == 0.9 | phi_selfpres_bin == 0.7) & phi_S1 == 0.9  ~ "look informative", 
      phi_informative_bin == 0.1 & phi_social_bin == 0.1 & phi_selfpres_bin == 0.9 & phi_S1 == 0.1 ~ "look social", 
      phi_informative_bin == 0.1 & phi_social_bin == 0.1 & phi_selfpres_bin == 0.9 & (phi_S1 == 0.3) ~ "look both",
      TRUE ~ "NA"))) %>%
  mutate(goal = factor(goal, levels = c("be truly informative", "be truly social", "be truly both", "look informative","look social", "look both","NA")))
```

#### schematic production

```{r}
fig.marginal.utility.s2 <- rs.s2 %>%
  mutate(utility = case_when(
    phi_informative > 0.9 ~ "informational",
    phi_social > 0.9 ~ "social",
    phi_selfpres > 0.9 & phi_S1 == 0.25 ~ "presentational \n(both-goal)",
    phi_selfpres > 0.9 & phi_S1 == 0.05 ~ "presentational \n(social)",
    phi_selfpres > 0.9 & phi_S1 == 0.95 ~ "presentational \n(informative)",
    phi_informative == phi_social & phi_selfpres < 0.1 ~ "inf, soc",
    phi_informative == phi_selfpres & phi_social < 0.1  & phi_S1 == 0.25~ "inf, pres",
    phi_social == phi_selfpres & phi_informative < 0.1  & phi_S1 == 0.25~ "soc, pres",
    phi_informative == phi_social & phi_social == phi_selfpres  & phi_S1 == 0.25~ "inf, soc, pres",
    TRUE ~ "NA"
  )) %>%
  mutate(utility = fct_relevel(utility, "informational", "social", "presentational \n(informative)", "presentational \n(social)", "presentational \n(both-goal)", "inf, soc", "inf, pres", "soc, pres", "inf, soc, pres")) %>%
  filter(state == 0, utility != "NA") %>%
  filter(utility %in% c("informational", "social", "presentational \n(informative)", "presentational \n(social)", "presentational \n(both-goal)")) %>%
  # filter(state == 0, phi_S1 %in% c(0.05, 0.25, 0.5, 0.75, 0.95), phi_selfpres > 0.9) %>%
  select(-phi_social, -phi_selfpres, -phi_informative, -state, -phi_S1) %>%
  gather(utt, prob, -utility) %>%
  separate(utt, into = c("positive", "utterance")) %>%
  mutate(positive = factor(positive, levels = c("yes", "not"),
                           labels = c("it was ~", "it wasn't ~")),
         utterance = factor(utterance, levels = c("terrible", "bad", "good", "amazing"))) %>%
  group_by(utility, positive, utterance) %>%
  summarize(prob = mean(prob)) %>%
  ggplot(., aes( x = utterance, color = positive, y = prob, group = positive))+
  geom_line()+
  facet_wrap(~utility, nrow = 1)+
  scale_color_ptol() +
  theme_few()+
  theme(axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1),
        strip.text.y = element_text(angle = 180),
        legend.title = element_blank(),
        legend.position = "bottom",
        title = element_text(size = 8),
        axis.title.y = element_text(size = 11))+
  # scale_y_continuous(limits = c(0, 0.75), breaks = c(0, 0.75), position = "left")+
  # ggtitle("Self-presentational mixture (0 = fully social; 1 = fully informational)")+
  ylab(expression("Schematic " * S[2] * "'s production probability"))
fig.marginal.utility.s2
ggsave("../../04_manuscript/01_general/fig/s2_schematic_production.png", width = 9, height = 4)
```

```{r fig.width=8}
cowplot::plot_grid(
  cowplot::plot_grid(
    fig.marginal.state.l1,
    fig.marginal.phi.l1,
    nrow = 1,
    rel_widths = c(1, 1.2),
    labels = c("A", "B")
  ),
  fig.marginal.self_pres.s2,
  nrow = 2, 
  rel_heights = c(1.1, 1),
  labels = c("", "D")
)

ggsave("../../04_manuscript/04_pnas/fig/L1_inferences_wS2pres.png", width = 9, height = 7)
```

#### S2 static

```{r}

rs.s2.tidy2 <- rs.s2.tidy %>%
                   mutate(utt2 = utt) %>%
                   mutate(utt = fct_recode(utt, "not amazing" = "not_amazing", 
                                            "not terrible" = "not_terrible", 
                                            "not good" = "not_good", 
                                            "not bad" = "not_bad", 
                                            "amazing" = "yes_amazing", 
                                            "good" = "yes_good", 
                                            "bad" = "yes_bad", 
                                            "terrible" = "yes_terrible"),
                          utt = fct_relevel(utt, "terrible", "bad", "good", "amazing", "not terrible", "not bad", "not good", "not amazing")) %>%
                            
                   separate(utt2, c("positive", "word")) %>%
                   mutate(word = fct_relevel(word, "terrible", "bad", "good", "amazing")) %>%
                   mutate(positive = fct_relevel(positive, "yes")) %>%
                   mutate(positive = fct_recode(positive, "it was ~" = "yes", "it wasn't ~" = "not")) %>%
                   mutate(goal = case_when(phi_informative_bin == 0.9 ~ "informational", 
                                           phi_social_bin == 0.9 ~ "social",
                                           phi_selfpres_bin == 0.9 & phi_S1 == 0.05 ~ "presentational \n(social S1)",
                                           phi_selfpres_bin == 0.9 & phi_S1 == 0.25 ~ "presentational \n(both-goal S1)",
                                           phi_selfpres_bin == 0.9 & phi_S1 == 0.95 ~ "presentational \n(informative S1)",
                                           # phi_social_bin == 0.5 & phi_informative_bin == 0.5 ~ "informational \n and social",
                                          phi_social_bin == 0.3 & phi_informative_bin == 0.3 & phi_selfpres_bin == 0.3 & phi_S1 == 0.25 ~ "informational, social \n and presentational")
                          ) %>%
                   mutate(goal = fct_relevel(goal, "informational", "social", "presentational \n(informative S1)", "presentational \n(social S1)", "presentational \n(both-goal S1)")) %>%
                   filter(!is.na(goal))
                   # filter(phi_informative_bin == 0.9)
```


```{r ms-fig-s2}
fig.s2 <- ggplot(rs.s2.tidy2, 
                 aes ( x = state, y = utt, fill = prob))+
  geom_tile(colour = "grey95") + 
  # scale_fill_viridis(option="plasma") +
  scale_fill_gradient(low = "white", high = "red")+
  facet_grid(positive ~ goal, switch="y", scales = "free_y")+
  # scale_y_continuous(limits = c(0, 1), breaks = c(0, 0.5, 1))+
  guides(fill = F)+
  ylab("utterance") +
  # scale_y_continuous(position = "right") +
  # scale_x_discrete(position = "top") +

  xlab("true state") +
  ggtitle(expression(S[2])) +
  
  # theme_bw()+
  theme_few() +
  theme(
    # axis.text.x = element_text(angle = 45, hjust=1),
        # strip.text.x = element_text(angle = 180),
        strip.text.y = element_blank()
        )  
# theme(strip.placement = "outside")
  # theme(        panel.border = element_rect(colour = "black", fill=NA, size=1))
  # panel_border(remove = TRUE)
  
fig.s2
ggsave("../../04_manuscript/01_general/fig/s2_utility_schematic_stateBottom.png", width = 10, height = 4)
```

```{r fig.width=8}
cowplot::plot_grid(
  # cowplot::plot_grid(
  #   fig.marginal.state.l1,
  #   fig.marginal.phi.l1,
  #   nrow = 1,
  #   rel_widths = c(1, 1.2),
  #   labels = c("A", "B")
  # ),
cowplot::plot_grid(
  fig.s2,
  nrow = 1, 
  labels = c("", "A")),  
cowplot::plot_grid(
  fig.marginal.utility.s2,
  nrow = 1, 
    rel_widths = c(0.2, 1.2),
  labels = c("", "B")), ncol=1 
)

ggsave("../../04_manuscript/01_general/fig/L1_S2_schematic.png", width = 9, height = 7)
```
